src/strerr.rs: Consolidate from Corroded `strerr_*.rs`
Consolidates all Corrode-translated source code for strerr-related
functionality into a common `strerr` module.

Moves all `strerr_*()` into an `impl strerr` block, as these functions
should eventually all accept `self`.

Code is otherwise unchanged from the original Corrode output.
